### 2024-02-13 #6349 @editorialbot
[PRE REVIEW]: DUGseis: A Python Package for Real-Time and Post-Processing of Picoseismicity
**Submitting author:** @mrosskopf (<a href="http://orcid.org/0009-0004-7156-4927">Martina Rosskopf</a>)
**Repository:** https://github.com/swiss-seismological-service/DUGseis
**Branch with paper.md** (empty if default branch):
**Version:** v0.3
