### 2024-03-01 963b0b5
Corrections on paper bib
