+++ paper/paper.bib
@@ -15,7 +15,7 @@
 
 @article{Greenfield2015, 
     author = {P Greenfield and M Droettboom and E Bray},
-    doi = {https://doi.org/10.1016/j.ascom.2015.06.004},
+    doi = {10.1016/j.ascom.2015.06.004},
     issn = {2213-1337},
     journal = {Astronomy and Computing},
     pages = {240-251},
@@ -48,7 +48,7 @@
 
 @misc{DugSeisAcqui,
     author = {Linus Villiger and Thomas Haag and Joseph Doetsch},
-    doi = {10.3929/ethz-b-000651477},
+    doi = {10.5281/zenodo.10728558},
     publisher = {ETH Research Collection},
     title = {DugSeis Acquisition},
     url = {https://github.com/swiss-seismological-service/DUGseis-acquisition/tree/main},
@@ -99,6 +99,7 @@
 @article{Villiger2020,
   title={Influence of reservoir geology on seismic response during decameter-scale hydraulic stimulations in crystalline rock},
   author={Villiger, L and Gischig, VS and Doetsch, J and Krietsch, H and Dutler, NO and Jalali, M and Valley, B and Selvadurai, PA and Mignan, A and Plenkers, K and others},
+  doi = {10.5194/se-11-627-2020},
   year={2020},
   journal = {Solid Earth},
   volume = {11},
