### 2024-04-08 7f20362
updated pyside6 version
