### 2024-04-15 #6349 @calum-chamberlain
> 👋 @calum-chamberlain - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sorry, but I am flat out with teaching at the moment and won't be able to provide a timely review.
