### 2024-04-30 #6349 @fwalter
Hello Chris,

I cannot see the paper I am supposed to review.

Fabian.

________________________________
Von: Chris Vernon ***@***.***>
Gesendet: Freitag, 26. April 2024 14:16
An: openjournals/joss-reviews ***@***.***>
Cc: Fabian Walter ***@***.***>; Mention ***@***.***>
Betreff: Re: [openjournals/joss-reviews] [PRE REVIEW]: DUGseis: A Python Package for Real-Time and Post-Processing of Picoseismicity (Issue #6349)

👋 @nialov<https://github.com/nialov> - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
