### 2024-05-03 #6349 @fwalter
I will have to decline, sorry for the late response. Several of the authors are part of an ongoing research collaboration and one may appear on a paper with me soon. I am not comfortable with this review.

Fabian.

Von: Chris Vernon ***@***.***>
Datum: Freitag, 3. Mai 2024 um 13:53
An: openjournals/joss-reviews ***@***.***>
Cc: Fabian Walter ***@***.***>, Mention ***@***.***>
Betreff: Re: [openjournals/joss-reviews] [PRE REVIEW]: DUGseis: A Python Package for Real-Time and Post-Processing of Picoseismicity (Issue #6349)

👋 @SaumikDana<https://github.com/SaumikDana> - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
