### 2024-05-21 #6768 @crvernon
👋 @mrosskopf and @ThomasLecocq - This is the review thread for the paper. All of our communications will happen here from now on.
