### 2024-05-22 #6768 @trichter
I have a conflict of interest with the last author of this submission. We are both long-time contributors to the [obspy package](https://github.com/obspy/obspy/graphs/contributors). I request this COI to be waived.
