### 2024-05-23 swiss-seismological-service/DUGseis#6 @trichter
improve first-time user experience with example dataset
Link to review issue: https://github.com/openjournals/joss-reviews/issues/6768

Hi!

As part of the JOSS review I tried to test DUGseis with the example dataset hosted at zenodo.

It would be nice if this example dataset was mentioned not only in the readme but also in the documentation.
I had to change the paths in the configuration file. Maybe these could be adapted in an update of the zenodo repository, or you could give instructions in the documenation.

After that, dugseis started without problems. But I have problems loading the waveforms. When the window opens, no waveforms are displayed.

![dugseisgui](https://github.com/swiss-seismological-service/DUGseis/assets/2046150/860ca892-095b-43fb-87e0-fb51d0bfcad9)

The error in the screenshot is because I clicked on "plot all traces", it has nothing to do with the above problem. However, the following warning might be relevant:

```
~/tmp/dugdata$ dug-seis gui --operator Tom --config post_processing_PaperExample.yaml
Creating/updating cache: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20/20 [00:00<00:00, 902.18it/s]
/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/OpenGL/arrays/numpymodule.py:200: RuntimeWarning: overflow encountered in cast
  return numpy.ascontiguousarray( source, typeCode )
Creating/updating cache: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20/20 [00:00<00:00, 51590.46it/s]

```

Any ideas?
