### 2024-05-23 #6768 @crvernon
@trichter - as long as you and the author have not collaborated on this current submission, I believe it is OK to waive this COI since the "obspy" package has so many contributors.
