### 2024-05-28 deba783
updated pyqtgraph version, removed command in README.md
