### 2024-05-28 swiss-seismological-service/DUGseis#6 @mrosskopf
Hi!

Thank you! Yes, it is a very good point to mention the example dataset in the documentation. I will add it in the following days and give instructions on what to change to use it.

This was probably related to the pygtgraph version. I updated the version in the installation .toml file. Could you try and upgrade pyqtgraph with "pip install --upgrade pyqtgraph" and then rerun the dug-seis gui command? It should hopefully solve the issue. If it works the traces that you select should then display in the left black window and if you click on plot all traces a new window should open with all traces.

Let me know if it works. :)
