### 2024-05-28 9afbf27
Force a deterministic order of event origins.
