### 2024-05-29 4c5b73f
Add readthedocs config.
