### 2024-06-21 swiss-seismological-service/DUGseis#6 @trichter
Thanks! With the updated documentation I got it working.
At first, I accidentally started the gui before running the processing script, which caused an error:

```
Traceback (most recent call last):
  File "/home/eule/dev/other/DUGseis/dug_seis/graphical_interface/main.py", line 1229, in on_channel_list_widget_itemSelectionChanged
    self._show_channels(selected_channels)
  File "/home/eule/dev/other/DUGseis/dug_seis/graphical_interface/main.py", line 456, in _show_channels
    plot = self.ui.plotWidget.addPlot(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/pyqtgraph/graphicsItems/GraphicsLayout.py", line 71, in addPlot
    plot = PlotItem(**kargs)
           ^^^^^^^^^^^^^^^^^
  File "/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/pyqtgraph/graphicsItems/PlotItem/PlotItem.py", line 159, in __init__
    self.titleLabel = LabelItem('', size='11pt', parent=self)
                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/pyqtgraph/graphicsItems/LabelItem.py", line 20, in __init__
    GraphicsWidget.__init__(self, parent)
  File "/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/pyqtgraph/graphicsItems/GraphicsWidget.py", line 17, in __init__
    QtGui.QGraphicsWidget.__init__(self, *args, **kargs)
  File "/home/eule/.local/mambaforge/envs/dug_seis/lib/python3.11/site-packages/pyqtgraph/graphicsItems/GraphicsWidgetAnchor.py", line 21, in __init__
    self.geometryChanged.connect(self.__geometryChanged)
RuntimeError: '__init__' method of object's base class (LabelItem) not called.
```

I understand that the gui is only for refinement and this is properly described in the documentation. So no urgent need to fix this, but there could be a better error message.

When I run the python processing script from the example I see some lines like

```
Possible error
Possible error
Possible error
```

The `print("Possible error")` in `amplitude_based_magnitudes.py` looks like a debug leftover. Also the comment in line 91/92 does not describe what the code actually does. Please check.

After that I can start the gui and everything works fine.
