### 2024-06-21 #6768 @trichter
@crvernon Sorry for the delay, I had to find a long enough time slot for the review.

I've finished my review. The submission looks good to me.

I have two minor issues:
  * Line 47: Please clarify the format of the station metadata, on first reading I thought it had to be in ASDF format.
  * I checked the Community Guidelines box because IMO it is pretty clear how to contribute or ask for help (open issues, create PRs, e.t.c), but technically this box asks for more. Maybe add a section in the readme?

My other comments have been addressed in the issues. The new points in the open issue are minor.
