### 2024-07-08 ff624e2
Small correction in the paper and added section in readme file how to get help.
