+++ paper/paper.md
@@ -55,7 +55,7 @@ The open source, Python-based DUGseis package is designed to align with the func
 
 # Functionality and Features
 The DUGseis software is a Python-based package with the main focus to process continuous high-frequency data, extract picoseismic event waveforms and create an earthquake catalog. To make its usage and its outputs more easily accessible for seismologists, some features use modules and functions of the ObsPy package [@Beyreuther2010], a popular package in seismology.\
-The DUGseis software requires sensor metadata with sensor locations and continuous waveform data in the ASDF format [@Greenfield2015; @asdf] to run. Here, the continuous waveform data was acquired using  DUGseis acquisition [@DugSeisAcqui], another software package developed specifically for the Bedretto Underground Laboratory to record and store waveform data of AE sensors in the ASDF format using specific Spectrum digitizer cards. For general usage of the here presented DUGseis package, the waveform data does not need to be acquired from DUGseis acquisition but the ASDF format can be transferred from other data formats. \
+The DUGseis software requires sensor metadata with sensor locations as StationXML files and continuous waveform data in the ASDF format [@Greenfield2015; @asdf] to run. Here, the continuous waveform data was acquired using  DUGseis acquisition [@DugSeisAcqui], another software package developed specifically for the Bedretto Underground Laboratory to record and store waveform data of AE sensors in the ASDF format using specific Spectrum digitizer cards. For general usage of the here presented DUGseis package, the waveform data does not need to be acquired from DUGseis acquisition but the ASDF format can be transferred from other data formats. \
 DUGseis retrieves all important information regarding data directories and processing settings from a configuration file. The configuration file can be used to open a graphical interface or can be given to the processing script. After processing the continuous waveform data, DUGseis outputs an event catalog as a database, which additionally can be saved as QuakeML or CSV files. Figure \autoref{fig:DUGseisScatch} shows the in- and output of the DUGseis processing.
 
 ![In- and output of DUGseis processing. \label{fig:DUGseisScatch}](DUGseis_InputOutputScatch.png)
