### 2024-07-08 #6768 @mrosskopf
Thank you @trichter for your review.
I adjusted the two minor issues in the paper and readme. I will also work on the suggestions you gave in the open issue in the next days.
