### 2024-07-17 #6768 @crvernon
Great @jessepisel! How about sometime within the next two weeks.  You can generate your review checklist using:

`@editorialbot generate my checklist`

Thanks so much!
