### 2024-08-09 #6768 @mrosskopf
Hello @crvernon,
sadly, this review is not really moving forward. Is there a way to speed it up somehow? This paper will be part of my PhD so it would be great to have it done soonish since it was published already beginning of February.
Until it is published, is there a way to cite this paper, because it does not have a DOI yet? It would be needed since other papers submitted later were already accepted.
Thanks already for your efforts!
