### 2024-08-10 #6768 @crvernon
Unfortunately @mrosskopf, we are bound by the time by which the reviewers can complete their reviews.   Since this paper has not yet been accepted for publication, the only citation that can be issued is the following one without a formal DOI:

> Rosskopf et al. (in review). DUGseis: A Python package for real-time and post-processing of picoseismicity. Journal of Open Source Software, 0(0), 4 6768.

I can move forward with my portion of this review once we have at least 2 reviewers sign off - having completed all of their checkboxes.  To that measure...

@trichter, @erexer, @jessepisel please provide an update as to when you will be able to finish your review of this work so that we may move it forward.
