### 2024-08-12 #6768 @erexer
@mrosskopf can you say more about the author list? I see Martina Rosskopf, Virginie Durand, Linus Villiger, Joseph Doetsch, Anne Obermann, and Lion Krischer listed as authors on the paper, and only four of those folks listed as contributors in the repo. (not sure who `seismotologist` is)
