### 2024-08-16 #6768 @mrosskopf
@erexer - Sure, I am guessing the four you are referring to are Virginie Durand, Linus Villiger, Lion Krischer and me. We worked together on bringing DUGseis to the state where it is now. Anne Obermann is a seismologist and my PhD supervisor, so she gave input on the software itself and helped with the paper. Joseph Doetsch was a main contributor on an earlier version of DUGseis which was used as a starting point for the current version.
I hope I answered your question. Let me know if there is anything else.
