### 2024-08-16 8943953
added small description in the amplitude function and removed debug leftover
