### 2024-08-27 3ab60d2
added suggestions by flake8 and black command in several files
