+++ dug_seis/event_processing/magnitudes/amplitude_based_magnitudes.py
@@ -40,7 +40,7 @@ def is_time_between(begin_time, end_time, check_time):
 
 def amplitude_based_relative_magnitude(st_event, event):
     # main parameters for magnitude processing
-    V_S = 3100.  # [m/s]
+    V_S = 3100.0  # [m/s]
     filter_freq_min = 3e3  # [Hz]
     filter_freq_max = 12e3  # [Hz]
     filter_corners = 4
@@ -48,8 +48,8 @@ def amplitude_based_relative_magnitude(st_event, event):
     signal_copy = st_event.copy()
     noise_copy = st_event.copy()
     gainLogAE = 100  # 10dB + 30dB
-    gainAE = 1. / gainLogAE
-    Count2VoltAE = 10. / 32000  # 10V on 32000 samples
+    gainAE = 1.0 / gainLogAE
+    Count2VoltAE = 10.0 / 32000  # 10V on 32000 samples
     conversion_factor_counts_mV = gainAE * Count2VoltAE  # assuming 64'000 (not 2**16 = 65536) counts
     # resolution per +/-10V (direct communication by GMuG) as well as 30dB pre-amplification
     # and 10 dB amplification from the supply/filter unit
@@ -58,23 +58,31 @@ def amplitude_based_relative_magnitude(st_event, event):
     Q = 76.0  # Quality factor [] introduced by Hansruedi Maurer (Email 29.07.2021)
     V_P = 5100.0  # P-wave velocity [m/s]
     r_0 = 10.0  # reference distance [m]
-    ### For amplitude
+    # For amplitude
     p_amp = []
     n_amp = []
     t_window = []
     distances = []
-    ### For magnitude
+    # For magnitude
     s_m = []
     Mr_station = []
 
     count = 0
     tmpPicks = event.picks
-    if len(tmpPicks)>len(event.preferred_origin().arrivals):
-        idx = [i for i,x in enumerate(tmpPicks) if x.evaluation_mode=='manual']
+    if len(tmpPicks) > len(event.preferred_origin().arrivals):
+        idx = [i for i, x in enumerate(tmpPicks) if x.evaluation_mode == "manual"]
         PicksManual = [tmpPicks[x].waveform_id for x in idx]
-        idx2 = [i for i,x in enumerate(tmpPicks) if x.waveform_id in PicksManual and x.evaluation_mode=='automatic']
+        idx2 = [
+            i
+            for i, x in enumerate(tmpPicks)
+            if x.waveform_id in PicksManual and x.evaluation_mode == "automatic"
+        ]
         PickDouble = [tmpPicks[x].waveform_id for x in idx2]
-        Picks = [x for x in tmpPicks if x.waveform_id not in PickDouble or x.evaluation_mode=='manual']
+        Picks = [
+            x
+            for x in tmpPicks
+            if x.waveform_id not in PickDouble or x.evaluation_mode == "manual"
+        ]
     else:
         Picks = tmpPicks
 
@@ -85,7 +93,7 @@ def amplitude_based_relative_magnitude(st_event, event):
         delta_p_s = dist / V_S - dist / V_P
         signal_window_start_time = pick.time - 0.5 * delta_p_s
         signal_window_end_time = pick.time + delta_p_s
-        noise_window_start_time = pick.time - 2.5*delta_p_s
+        noise_window_start_time = pick.time - 2.5 * delta_p_s
         noise_window_end_time = pick.time - 0.5 * delta_p_s
 
         # skip magnitude computation if
@@ -115,7 +123,7 @@ def amplitude_based_relative_magnitude(st_event, event):
                 noise_window_end_time,
             )
         ):
-            #print('Possible error')
+            # print('Possible error')
             continue
 
         distances.append(
[…]
