+++ dug_seis/event_processing/picking/pickers/PhasePApy_Austin_Holland/ktpicker.py
@@ -18,7 +18,6 @@ from .cf_kt import *
 
 
 class KTPicker:
-
     """
     KTPicker is designed based on kurtosis.
     """
@@ -35,7 +34,6 @@ class KTPicker:
         pol_coeff=10,
         uncert_coeff=3,
     ):
-
         """
         Parameter description:
           t_win        : the time in seconds of moving window to calculate kurtosis
