+++ scripts/run_processing_incl_REC_pick_refinement.py
@@ -1,6 +1,7 @@
 """
 Example processing script.
 """
+
 import logging
 
 import obspy
