### 2024-08-27 81c400f
updated development section in the documentation
