### 2024-08-27 #6768 @jessepisel
Hi @crvernon I got the package installed and up and running this past week. I opened a couple of minor issues that the authors have fixed. I have completed my review and checklist at this time. It would be good to see more test coverage for the package at some point down the line, but the current tests have sufficient coverage. It looks like the only bit to sort out is the additional contributions from other authors raised earlier in the review process. Once that is complete, this is all done on my side.
