### 2024-09-04 swiss-seismological-service/DUGseis#11 @erexer
JOSS Review: Paper questions
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768).

- [x] Can you provide a source for the data used in Figure 2.
- [x] You mention that there are other programs used for seismic processing, can you give the names of those and cite them.

typos/grammar

- [x] "Figure Figure 2 shows the layout..." says "Figure" twice
- [x] "Other commercial software for monitoring picoseismicity for structural health monitoring applications, e.g. in mines exists, but are only trigger-based" the commas in this sentence are incorrect, should be "Other commercial software for monitoring picoseismicity for structural health monitoring applications, e.g., in mines, exists but are only trigger-based".
- [x] "For large-scale experiments in underground laboratories (e.g., hydraulic stimulation, earthquake nucleation, nuclear waste disposal) continuous recordings of the seismicity data streams in MHz range are needed to study the rock response in great detail" should have a comma "For large-scale experiments in underground laboratories (e.g., hydraulic stimulation, earthquake nucleation, nuclear waste disposal), continuous recordings of the seismicity data streams in MHz range are needed to study the rock response in great detail"
- [x] "Since the software is python based, users can easily add their own processing routines." Python should be capitalized.

I'd recommend running the paper through a grammar checker.
