### 2024-09-04 swiss-seismological-service/DUGseis#12 @erexer
JOSS Review: tests
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768).

I'm running the tests for `DUGseis` with Python 3.12.5 and got a lot of failed tests, running the command `$ py.test` in the `DUGseis` dir.

```bash
========================================= short test summary info ==========================================
FAILED dug_seis/tests/test_db.py::test_reading_and_writing_full_pick - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_retrieving_pick_by_resource_id - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_reading_and_writing_single_comment - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_reading_and_writing_full_magnitude - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_reading_and_writing_full_origin - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_reading_and_writing_arrival - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_read_and_write_focal_mechanism - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files[iris_events.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files[neries_events.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files[usgs_event.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files[event_with_preferred_ids.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files[quakeml_1.2_event.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_save_as_quakeml - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_event_to_disc - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_get_event_by_resource_id - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files_and_delete_whole_files[iris_events.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files_and_delete_whole_files[neries_events.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files_and_delete_whole_files[usgs_event.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files_and_delete_whole_files[event_with_preferred_ids.xml] - TypeError: attribute id needs to be a string.
FAILED dug_seis/tests/test_db.py::test_round_trip_whole_files_and_delete_whole_files[quakeml_1.2_event.xml] - TypeError: attribute id needs to be a string.
===================== 20 failed, 73 passed, 2 xfailed, 1 warning in 109.04s (0:01:49) ======================
```

here's the results of `pip list`

``` bash
Package            Version     Editable project location
------------------ ----------- ------------------------------------------------------------------------
certifi            2024.8.30
charset-normalizer 3.3.2
click              8.1.7
colorama           0.4.6
contourpy          1.3.0
cycler             0.12.1
decorator          5.1.1
dill               0.3.8
DUG-Seis           0.3         /Users/<me>/Library/CloudStorage/OneDrive/Documents/JOSS/DUGseis
fonttools          4.53.1
greenlet           3.0.3
h5py               3.11.0
idna               3.8
iniconfig          2.0.0
isodate            0.6.1
joblib             1.4.2
kiwisolver         1.4.7
llvmlite           0.43.0
lxml               5.3.0
matplotlib         3.9.2
networkx           3.3
numba              0.60.0
numpy              2.0.2
obspy              1.4.1
packaging          24.1
pillow             10.4.0
pip                24.2
pluggy             1.5.0
prov               2.0.1
pyasdf             0.8.1
PyOpenGL           3.1.7
pyparsing          3.1.4
pyproj             3.6.1
pyqtgraph          0.13.7
PySide6            6.6.1
PySide6-Addons     6.6.1
PySide6-Essentials 6.6.1
pytest             8.3.2
python-dateutil    2.9.0.post0
PyYAML             6.0.2
rdflib             6.3.2
requests           2.32.3
schema             0.7.7
scipy              1.14.1
setuptools         74.1.1
shiboken6          6.6.1
six                1.16.0
SQLAlchemy         1.4.53
tqdm               4.66.5
urllib3            2.2.2
```

any idea why these are failing or if there's something wrong with my env?
