### 2024-09-06 2896f7f
updated paper text, Figure 2 and added database and yaml for Figure 2
