+++ paper/GUI_exampleEvent_Fig2.yaml
@@ -1,0 +1,69 @@
+version: 14
+
+meta:
+  project_name: VALTER_Stimu3
+  project_location: Bedretto
+  project_description: VALTER_Stimu3
+
+local_coordinate_system:
+  epsg_code: 2056
+  translation_vector: [2679720.696, 1151600.128, 0.0]
+
+paths: # absolute paths
+  asdf_folders:
+    - '/PATH/TO/WAVEFORMFOLDER/.../WAVEFORMS/continuous-04-bedretto'
+    - '/PATH/TO/WAVEFORMFOLDER/.../WAVEFORMS/continuous-03-bedretto'
+  stationxml_folders:
+    - '/PATH/TO/STATIONXMLFOLDER/.../STATIONXMLS'
+  database: 'sqlite:///PATH/TO/DATABASEOUTPUT/.../GUI_ExampleEvent_Fig2.sqlite'
+  cache_folder: '/PATH/TO/CACHEFOLDER/.../cache'
+
+# Must cover the whole range of the experiment, e.g. more than the expected
+# start + end time for the live test case.
+temporal_range:
+  start_time: 2023-03-14T15:50:00.000Z
+  end_time: 2023-03-14T15:57:00.000Z
+
+
+graphical_interface:
+  classifications:
+    - passive
+    - relocated
+    - noise
+    - unknown
+    - active
+  pick_types:
+    - P
+    - S
+  uncertainties_in_ms:
+    - 0.00001
+    - 0.000025
+    - 0.00005
+  3d_view:
+    # Helpful to distinguish newer from older events.
+    color_events: plasma
+    size_events_in_pixel: 7
+    size_channels_in_pixel: 3
+    # Greenish color for channels.
+    color_channels: [0.1, 0.9, 0.0, 0.4]
+
+  location_algorithm_default_args:
+    velocity:
+      P: 5100.0
+    damping: 0.01
+    use_anisotropy: true
+    anisotropy_parameters:
+      P:
+        azi: 310.0
+        inc: 28.6
+        delta: 0.071
+        epsilon: 0.067
+
+filters:
+  - filter_id: smi:local/bandpass_causal_1000_5000
+    filter_settings:
+      filter_type: butterworth_bandpass
+      highpass_frequency_in_hz: 1000.0
+      lowpass_frequency_in_hz: 5000.0
+      filter_corners: 4
+      zerophase: false
