### 2024-09-06 swiss-seismological-service/DUGseis#11 @mrosskopf
Hi,
thank you for your suggestions. I adjusted the text and also ran it through a grammar checker again.
I changed the event that was displayed in Figure 2. The one displayed now is just a few hours later, but the waveforms and channel information are already on zenodo. To open it in GUI, you can use the yaml file (just change the paths) and the sqlite file I now provided in the paper folder of the repository, plus the given waveform and channel information from zenodo.
Are there any other suggestions to this topic? :)
