### 2024-09-06 d75cb9a
removed sources (brackets in a function and wrong variable name) for failed tests
