### 2024-09-06 swiss-seismological-service/DUGseis#12 @mrosskopf
There is nothing wrong with your environment. I did some changes after running the black command last week, and it seems that one suggestion influenced the code in a wrong way. I changed it back now. You can update the repo on your PC and then the fails should be gone. Let me know if it is not the case.
