### 2024-09-09 swiss-seismological-service/DUGseis#12 @erexer
great, that fixed it!
