### 2024-09-09 swiss-seismological-service/DUGseis#11 @erexer
Thanks! Could you add GMuG as a citation and make it clear in Figure 2 that the data comes from Rosskopf et al. 2024 (put it in the caption).
