### 2024-09-16 1e9b35a
updated paper - added software citation
