+++ paper/paper.bib
@@ -67,7 +67,21 @@
     volume = {81},
     url = {https://doi.org/10.1785/gssrl.81.3.530},
     year = {2010},
-}  
+}
+
+@inbook{Plenkers2022,
+   abstract = {In-situ acoustic emission (AE) monitoring is carried out in mines, tunnels and underground laboratories in the context of structural health monitoring, in decameter-scale research projects investigating the physics of earthquake nucleation and propagation and in research projects looking into the seismo-hydro-mechanical response of the rock mass in the context of hydraulic stimulations or nuclear waste storage. In addition surface applications e.g. monitoring rock faces of large construction sites, rock fall areas and rock slopes are documented in the literature. In geomechanical investigations in-situ AE monitoring provides information regarding the stability of underground cavities, the state of stress and the integrity of the rock mass. The analysis of AE events recorded in-situ allows to bridge the observational gap between the studies of faulting processes in laboratory and studies of larger natural and induced earthquakes. This chapter provides an overview of various projects involving in-situ AE monitoring underground with a focus on recent achievements in the field. In-situ AE monitoring networks are able to record AE activity from distances up to 200 m, but the monitoring limits depend strongly on the extension of the network, geological and tectonic conditions. Very small seismic events with source sizes on approximately decimeter to millimeter scale are detected. In conclusion in-situ AE monitoring is a useful tool to observe instabilities in rock long before any damage becomes directly visible and is indispensable in high-resolution observations of rock volume deformation in decameter in-situ rock experiments.},
+   author = {Katrin Plenkers and Gerd Manthei and Grzegorz Kwiatek},
+   doi = {10.1007/978-3-030-67936-1_16},
+   issn = {23662603},
+   journal = {Springer Tracts in Civil Engineering},
+   keywords = {Earthquake physics,Hydraulic stimulation,In-situ AE monitoring,Rock stability,Scaling relations,Structural health monitoring,Underground in-situ experiments},
+   pages = {403-476},
+   publisher = {Springer Science and Business Media Deutschland GmbH},
+   title = {Underground In-situ Acoustic Emission in Study of Rock Stability and Earthquake Physics},
+   year = {2022},
+}
+
 
 @article{Plenkers2023,
     author = {Katrin Plenkers and Andreas Reinicke and Anne Obermann and Nima Gholizadeh Doonechaly and Hannes Krietsch and Thomas Fechner and Marian Hertrich and Karam Kontar and Hansruedi Maurer and Joachim Philipp and Beat Rinderknecht and Manuel Volksdorf and Domenico Giardini and Stefan Wiemer},
@@ -105,3 +119,4 @@
   volume = {11},
   pages = {627-655},
 }
+
