### 2024-09-16 swiss-seismological-service/DUGseis#11 @mrosskopf
Hi, I slightly changed the text to not include a citation in the summary section but I added the website of GMuG and of another software in the section "Statement of need" after slightly changing the text. The software itself do not have a direct citation, but I hope it is fine with the websites.
I also added the data citation in the caption.
