+++ paper/paper.bib
@@ -8,9 +8,10 @@
 
 @article{Obermann2024,
   	Author = {Anne Obermann and Martina Rosskopf and Virginie Durand and Katrin Plenkers and Kai Bröker and Nima Gholizadeh Doonechaly and Valentin Gischig and Marian Hertrich and Philipp Kästli and Xiaodong Ma and Hansruedi, Maurer and Antonio Pio Rinaldi and Linus Villiger and Stefan Wiemer and Alba Zappone and Domenico Giardini},
-  	Journal = {Submitted to Journal of Geophysical Research},
+  	Journal = {Journal of Geophysical Research (in review)},
   	Title = {{Picoseismic response of hectometer-scale fracture systems to stimulation with cm-scale resolution under the Swiss Alps, in the Bedretto Underground laboratory}},
   	Year = 2024
+  	url = {http://dx.doi.org/10.1029/2024JB029836}
 }
 
 @article{Greenfield2015, 
