### 2024-09-16 swiss-seismological-service/DUGseis#11 @erexer
You need a comma at the end of [this line](https://github.com/swiss-seismological-service/DUGseis/blob/main/paper/paper.bib#L13) so that the paper will render correctly :)
