### 2024-09-17 swiss-seismological-service/DUGseis#11 @mrosskopf
Ah yes, of course! Sorry about that.
