### 2024-09-30 swiss-seismological-service/DUGseis#13 @erexer
JOSS Review: numpy error
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768).

I'm running the GUI and while I'm running it, I'm getting errors.

```bash

A module that was compiled using NumPy 1.x cannot be run in
NumPy 2.0.2 as it may crash. To support both 1.x and 2.x
versions of NumPy, modules must be compiled with NumPy 2.0.
Some module may need to rebuild instead e.g. with 'pybind11>=2.12'.

If you are a user of the module, the easiest solution will be to
downgrade to 'numpy<2' or try to upgrade the affected module.
We expect that some modules will need time to support NumPy 2.
```

```bash
Traceback (most recent call last):
  File "/<myfilepath>/DUGseis/dug_seis/graphical_interface/main.py", line 823, in _reload_data_after_waveform_data
    self._init_3d_view(set_camera=False)
  File "/<myfilepath>/DUGseis/dug_seis/graphical_interface/main.py", line 198, in _init_3d_view
    self.three_d_view.update_channels(
  File "/<myfilepath>/DUGseis/dug_seis/graphical_interface/three_d_view.py", line 236, in update_channels
    bounds = self.compute_bounds(coordinates=channel_coordinates)
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/<myfilepath>/DUGseis/dug_seis/graphical_interface/three_d_view.py", line 309, in compute_bounds
    scale = max(coordinates.ptp(axis=0))
                ^^^^^^^^^^^^^^^
AttributeError: `ptp` was removed from the ndarray class in NumPy 2.0. Use np.ptp(arr, ...) instead.
```
