### 2024-09-30 swiss-seismological-service/DUGseis#14 @erexer
JOSS Review: advice for example 
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768), but isn't necessary for approval :)

The example has more steps than necessary to run out of the box.

1. Ideally, you'd have a download command, `wget "<url>"`, so that users don't need to manually go to the site to download the example data.

2. The [example dataset](https://dugseis.readthedocs.io/en/latest/example_dataset.html) has numerous steps to prepare the [download data](https://zenodo.org/records/10598393), it would be a much lower barrier of entry if you set up the example to run out of the box (organize the directory structure, create an empty cache dir, include an `.sqlite` file, edit the config file to work with the example data) so that all the user has to do is download the zip file, unzip the dir, `cd` into the dir, install the package, and run the `run_postProcessing_PaperExample.py` file.
