### 2024-10-03 bcb7f05
Changed package dependency for numpy to download numpy<2.0
