+++ pyproject.toml
@@ -39,7 +39,7 @@ dependencies = [
     "joblib",
     "matplotlib",
     "numba",
-    "numpy",
+    "numpy<2.0",
     "obspy",
     "pyasdf",
     "pyopengl",
