### 2024-10-03 a960298
Change dependency description in doc for numpy
