### 2024-10-03 swiss-seismological-service/DUGseis#13 @mrosskopf
Hi,
thank you! I changed the numpy version to smaller <2.0 since some packages don't work with the newest version yet. Could you downgrade your numpy e.g. to 1.26.4 and try again? With the smaller numpy version it should work.
Best!
