### 2024-10-03 00a008f
Updated the documentation of example dataset for easier usage with a zip folder uploaded to zenodo and changed DOI in paper bib file
