+++ paper/paper.bib
@@ -42,7 +42,7 @@
     month = {02},
     publisher = {Zenodo},
     title = {DUGseis processing example with needed data files},
-    url = {https://doi.org/10.5281/zenodo.10598393},
+    url = {https://doi.org/10.5281/zenodo.10598392},
     year = {2024},
 }
 
