+++ docs/example_dataset.md
@@ -1,19 +1,28 @@
 # Example dataset
 
 An example dataset consisting of waveform data, station information, processing and config file can be found here:
-[Zenodo - DUGSeis Example Dataset](https://doi.org/10.5281/zenodo.10598393)
+[Zenodo - DUGSeis Example Dataset](https://doi.org/10.5281/zenodo.10598392)
+
+
 
 ## Usage
+After downloading the zip file, extract the content in the zip folder. The folder `BedrettoExampleData` contains the following folders:
+* `WAVEFORMS`: containing the waveform files in asdf format
+* `STATIONXMLS`: containing the station information files in xml format
+* `cache`: the folder where the cache files will be saved (empty until first run of graphical user interface or post-processing script)
 
-After downloading all files, the easiest way to try out the example is to add the waveform files (.h5 files) to a seperate folder (e.g. WAVEFORMS) and the station information files (.xml files) to another seperate folder (e.g. STATIONXMLS).
-If the processing file (run_postProcessing_PaperExample.py) and the config file (post_processing_PaperExample.yaml) are located in the same folder, nothing needs to be changed in the processing file.
-In the config file the `paths` need to be updated:
-* `asdf_folders`: Absolute path to the folder containing the waveform files
-* `stationxml_folders`: Absolute path to the folder containing the station information files
-* `database`: Absolute path to where the database file should be saved including the name of the database file
-* `cache_folder`: Absolute path to the cache folder, which needs to be created before running the processing
+Additionally, the folder contains the following files:
+* `GUIExample_PaperExample.yaml`: the config file for quickly opening the graphical user interface
+* `VALTERStimu_PaperDatabase.sqlite`: the database file containing the results of the processing displayed in the graphical user interface
+* `post_processing_PaperExample.yaml`: the config file for the post-processing script
+* `run_postProcessing_PaperExample.py`: Python script to run the post-processing
 
-After updating the paths in the config file, the processing can be started by running the following command in the terminal:
+To open the graphical interface, run the following command in the terminal:
+```bash
+dug-seis gui --operator="Your Name" --config=GUIExample_PaperExample.yaml
+```
+
+To run the post-processing, run the following command in the terminal:
 ```bash
 python3 run_postProcessing_PaperExample.py
 ```
