### 2024-10-04 swiss-seismological-service/DUGseis#14 @mrosskopf
Hi,
thanks for the suggestion! I added a new version to zenodo with a zip file, where your point 2 is taken care of and updated the documentation for it as well.
I need to have a better look at your first point. For now, I could not download the data via the terminal.
Best
