### 2024-10-09 swiss-seismological-service/DUGseis#11 @erexer
Thanks for bearing through this with me! There's still some grammar errors in the paper, for example, when I run the abstract through [this grammar checker](https://languagetool.org/), it points out:
- "detection of _the_ tiniest" should include "the"
- "(e.g._,_ hydraulic stimulation," should have a comma after e.g.
- "developed to manage, process_,_ and visualize continuous" may need an Oxford comma
- "software to experiment _with_ specific questions" requires a preposition
