### 2024-10-09 swiss-seismological-service/DUGseis#15 @erexer
JOSS Review: documentation 
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768). Got some questions/comments.

- [x] [Installation](https://dugseis.readthedocs.io/en/latest/installation.html): "Furthermore git is required to get the DUGSeis code." git isn't required, the user can download the code by going to the [DUGseis repo](https://github.com/swiss-seismological-service/DUGseis), clicking `Code` and then selecting `Download ZIP`. You could say instead that git is an option for downloading the code.
- [x] (not required for JOSS acceptance) In [Linting](https://dugseis.readthedocs.io/en/latest/development.html#linting), you say to `pip install flake8 black`. It's cleaner to keep all the dependencies in one place, the `pyproject.toml`. The [guide to writing your pyproject.toml](https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#dependencies-optional-dependencies) guides you through creating optional dependencies so that you can keep your development dependencies separate from runtime dependencies.
- [x] (not required for JOSS acceptance) The above comment also applies to [Documentation Building](https://dugseis.readthedocs.io/en/latest/development.html#documentation-building).
- [x] [Graphical Interface](https://dugseis.readthedocs.io/en/latest/development.html#graphical-interface): the command listed assumes that the user is on a windows machine (I think?), the linux/mac command is `pyside6-uic ./main_window.ui -o main_window.py`. It would be good to list both.
- [x] [StationXML Meta Data](https://dugseis.readthedocs.io/en/latest/data_requirements.html#stationxml-meta-data): "Metadata, e.g._,_ location and instrument" should not have a space between "meta" and "data" and have a comma after e.g.

Love the development page, its super informative
