### 2024-10-09 swiss-seismological-service/DUGseis#16 @erexer
JOSS Review: installation fails for python 3.13.0
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/6768).

Installation fails for python 3.13.0

```bash
pip install -e .                                 ✔ │ 3.13.0rc3t  │ 10:37:20
Obtaining file:///<filepath>/DUGseis
  Installing build dependencies ... done
  Checking if build backend supports build_editable ... done
  Getting requirements to build editable ... done
  Preparing editable metadata (pyproject.toml) ... done
Collecting click (from DUG-Seis==0.3)
  Using cached click-8.1.7-py3-none-any.whl.metadata (3.0 kB)
Collecting joblib (from DUG-Seis==0.3)
  Using cached joblib-1.4.2-py3-none-any.whl.metadata (5.4 kB)
Collecting matplotlib (from DUG-Seis==0.3)
  Downloading matplotlib-3.9.2-cp313-cp313t-macosx_10_13_x86_64.whl.metadata (11 kB)
Collecting numba (from DUG-Seis==0.3)
  Downloading numba-0.60.0.tar.gz (2.7 MB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [24 lines of output]
      Traceback (most recent call last):
        File "/<filepath>/DUGseis/.venv3.13.0/lib/python3.13t/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
          ~~~~^^
        File "/<filepath>/DUGseis/.venv3.13.0/lib/python3.13t/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
                                   ~~~~^^^^^^^^^^^^^^^^^^^^^^^^
        File "/<filepath>/DUGseis/.venv3.13.0/lib/python3.13t/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 118, in get_requires_for_build_wheel
          return hook(config_settings)
        File "/private/var/folders/81/h441ylh90yn8p6cmrrlp17w00000gn/T/pip-build-env-9_tmulqo/overlay/lib/python3.13t/site-packages/setuptools/build_meta.py", line 332, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=[])
                 ~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/81/h441ylh90yn8p6cmrrlp17w00000gn/T/pip-build-env-9_tmulqo/overlay/lib/python3.13t/site-packages/setuptools/build_meta.py", line 302, in _get_build_requires
          self.run_setup()
          ~~~~~~~~~~~~~~^^
        File "/private/var/folders/81/h441ylh90yn8p6cmrrlp17w00000gn/T/pip-build-env-9_tmulqo/overlay/lib/python3.13t/site-packages/setuptools/build_meta.py", line 503, in run_setup
          super().run_setup(setup_script=setup_script)
          ~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "/private/var/folders/81/h441ylh90yn8p6cmrrlp17w00000gn/T/pip-build-env-9_tmulqo/overlay/lib/python3.13t/site-packages/setuptools/build_meta.py", line 318, in run_setup
          exec(code, locals())
          ~~~~^^^^^^^^^^^^^^^^
        File "<string>", line 51, in <module>
        File "<string>", line 48, in _guard_py_ver
      RuntimeError: Cannot install on Python version 3.13.0; only versions >=3.9,<3.13 are supported.
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```
