### 2024-10-09 swiss-seismological-service/DUGseis#13 @erexer
It's working great :)
