### 2024-10-10 swiss-seismological-service/DUGseis#11 @mrosskopf
I hope I corrected all errors now. I already used your suggested grammar checker in the past, but only the free version. I tried different ones again and hope they fix all the grammar mistakes.
