### 2024-10-10 swiss-seismological-service/DUGseis#16 @mrosskopf
The python version is too high for now. It works stable with python 3.11. Therefore, in the installation description, the python version for the conda environment is fixed to 3.11.
I also added it again to the Dependencies part in the documentation where I state now python=3.11 instead of >=3.9
I hope this is fine like that.
