### 2024-10-10 da0a9e8
Added optional packages to pytoml and changed documentation for installation of the optional packages.
