### 2024-10-10 swiss-seismological-service/DUGseis#15 @mrosskopf
Thanks for the comments. All suggestions are now added/changed in the documentation and the optional packages added to the pytoml.
