### 2024-10-11 d0d59c4
Updated paper and and bib file
