+++ paper/paper.bib
@@ -7,12 +7,13 @@
 }
 
 @article{Obermann2024,
-  	Author = {Anne Obermann and Martina Rosskopf and Virginie Durand and Katrin Plenkers and Kai Bröker and Nima Gholizadeh Doonechaly and Valentin Gischig and Marian Hertrich and Philipp Kästli and Xiaodong Ma and Hansruedi, Maurer and Antonio Pio Rinaldi and Linus Villiger and Stefan Wiemer and Alba Zappone and Domenico Giardini},
-  	Journal = {Journal of Geophysical Research (in review)},
-  	Title = {{Picoseismic response of hectometer-scale fracture systems to stimulation with cm-scale resolution under the Swiss Alps, in the Bedretto Underground laboratory}},
-  	Year = {2024},
-  	url = {http://dx.doi.org/10.1029/2024JB029836},
+   author = {Anne Obermann and Martina Rosskopf and Virginie Durand and Katrin Plenkers and Kai Bröker and Antonio Pio Rinaldi and Nima Gholizadeh Doonechaly and Valentin Gischig and Alba Zappone and Florian Amann and Massimo Cocco and Marian Hertrich and Mohammadreza Jalali and Jonas Junker and Philipp Kästli and Xiaodong Ma and Hansruedi Maurer and Men-Andrin Meier and Miriam Schwarz and Paul Selvadurai and Linus Villiger and Stefan Wiemer and Luca Dal Zilio and Domenico Giardini},
+   doi = {http://dx.doi.org/10.1029/2024JB029836},
+   journal = {JGR Solid Earth (in review)},
+   title = {Seismic responce of hectometer scale fracture system to hydraulic stimulation in the Bedretto Underground laboratory, Switzerland},
+   year = {2024},
 }
+
 
 @article{Greenfield2015, 
     author = {P Greenfield and M Droettboom and E Bray},
@@ -112,12 +113,18 @@
 } 
 
 @article{Villiger2020,
-  title={Influence of reservoir geology on seismic response during decameter-scale hydraulic stimulations in crystalline rock},
-  author={Villiger, L and Gischig, VS and Doetsch, J and Krietsch, H and Dutler, NO and Jalali, M and Valley, B and Selvadurai, PA and Mignan, A and Plenkers, K and others},
-  doi = {10.5194/se-11-627-2020},
-  year={2020},
-  journal = {Solid Earth},
-  volume = {11},
-  pages = {627-655},
[…]
