+++ paper/paper.md
@@ -48,49 +48,49 @@ bibliography: paper.bib
 
 
 # Summary
-Detecting earthquakes and compiling them to earthquake catalogs are fundamental tasks in seismology. 
-Acoustic emission sensors allow the detection of the tiniest so-called picoseismic events, representing fractures on the 
-cm or dm scale (-5<magnitude<0), even smaller seismic events are recorded in the laboratory. Picoseismic events have 
-corner frequencies of 1 kHz-100 kHz. They are typically recorded with sampling frequencies of 200 kHz to 1 MHz and cannot be 
-handled by standard seismic processing software that is developed for signals with frequencies f<500 Hz. Commercial software 
-for continuous recording and processing of picoseismicity exists (e.g., GMuG-conti, fastloc.REEL) and is successfully used 
-in scientific in-situ experiments (e.g., hydraulic stimulation, earthquake nucleation, nuclear waste disposal). However,
-commercial software comes with restrictions unfavorable for university-based in-situ experiments in underground 
-laboratories, where tailoring the software to experiment with specific questions or adding newly developed analysis 
-techniques are routine operations in order to advance the study of the rock response in great detail. 
-The open-source DUGseis software package fills this gap. It has been developed to manage, process, and visualize 
-continuous, high-frequency seismic data. The package can be used to create real-time and post-processing earthquake 
-catalogs and to directly visualize their event waveforms and locations in a graphical interface. As the software is 
-Python-based, users can easily add their own processing routines.
+Detecting earthquakes and compiling the events into earthquake catalogs are fundamental tasks in seismology. Acoustic 
+emission sensors allow for the detection of the tiniest so-called picoseismic events, representing fractures on the 
+cm or dm scale (-5<magnitude<0). Even smaller seismic events are recorded in the laboratory. Picoseismic events have 
+corner frequencies of 1 kHz - 100 kHz. They are typically recorded with sampling frequencies of 200 kHz to 1 MHz and 
+cannot be handled by standard seismic processing software that is developed for signals with frequencies f<500 Hz. 
+Commercial software for continuous recording and processing of picoseismicity exists (e.g., GMuG-conti, fastloc.REEL) 
+and is successfully used in scientific in-situ experiments (e.g., hydraulic stimulation, earthquake nucleation, 
+nuclear waste disposal). However, commercial software comes with restrictions that are undesirable for university-based
+in-situ experiments in underground laboratories; where tailoring the software to experiment with specific questions or
+adding newly developed analysis techniques are routine operations in order to advance the study of the rock response 
+in greater detail. The open-source DUGseis software package fills this gap. It has been developed to manage, process, 
+and visualize continuous, high-frequency seismic data. The package can be used to create real-time and post-processing
+earthquake catalogs and to directly visualize their event waveforms and locations in a graphical interface. Since the 
+software is Python-based, users can easily add their own processing routines.
 
 # Statement of need
 The open-source, Python-based DUGseis package is designed to align with the functionalities of 
 SeisComP [@seiscomp], a standard software used in microseismic large-scale earthquake processing (<500Hz, M>-0.5). 
-DUGseis is tailored for picoseismic events (-5<M<0) with much higher frequency ranges between 1 kHz and 100 kHz [@Plenkers2022], as recorded by 
+DUGseis is tailored to picoseismic events (-5<M<0) with much higher frequency ranges between 1 kHz and 100 kHz [@Plenkers2022], as recorded by 
 acoustic emission sensors (AE sensors). High-frequency seismic data processing is common in mining environments to 
 monitor tunnel stability and has become very popular in underground laboratories [@Plenkers2022]. To date, many projects 
-used commercial software (e.g., GMuG-conti from GMuG Gesellschaft für Materialprüfung und Geophysik, Bad Nauheim/Germany,
[…]
