### 2024-10-11 swiss-seismological-service/DUGseis#11 @mrosskopf
For your information - I just updated the paper again since it was now corrected by a professional proofreader.
