### 2024-10-11 swiss-seismological-service/DUGseis#16 @erexer
Ah, ok. You should also change the python version in the [`pyproject.toml`](https://github.com/swiss-seismological-service/DUGseis/blob/main/pyproject.toml#L23).
