### 2024-10-11 swiss-seismological-service/DUGseis#14 @erexer
Great! I followed the updated instructions in the [Example Dataset](https://dugseis.readthedocs.io/en/latest/example_dataset.html) page and the command `dug-seis gui --operator="Your Name" --config=GUIExample_PaperExample.yaml` is failing for me with:
```bash

Traceback (most recent call last):
  File "/<filepath>/DUGseis/.venv/lib/python3.11/site-packages/schema.py", line 483, in validate
    nvalue = Schema(
             ^^^^^^^
  File "/<filepath>/DUGseis/.venv/lib/python3.11/site-packages/schema.py", line 440, in validate
    return type(data)(o.validate(d, **kwargs) for d in data)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/<filepath>/DUGseis/.venv/lib/python3.11/site-packages/schema.py", line 440, in <genexpr>
    return type(data)(o.validate(d, **kwargs) for d in data)
                      ^^^^^^^^^^^^^^^^^^^^^^^
  File "/<filepath>/DUGseis/.venv/lib/python3.11/site-packages/schema.py", line 219, in validate
    raise SchemaError(
schema.SchemaError: Or(Use(<function _directory_exists at 0x12835cae0>)) did not validate 'C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto'
_directory_exists('C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto') raised ValueError("Directory 'C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto' does not exist.")
```

it's also throwing
```bash
line 489, in validate
    raise SchemaError(
schema.SchemaError: Key 'paths' error:
Key 'asdf_folders' error:
Or(Use(<function _directory_exists at 0x12835cae0>)) did not validate 'C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto'
_directory_exists('C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto') raised ValueError("Directory 'C:/Users/Martina/Documents/PhD/workspace/BedrettoProcessingServer/DUGseis_RUNS/DUGseis_Paper/WAVEFORMS/continuous-04-bedretto' does not exist.")
```
