### 2024-10-15 swiss-seismological-service/DUGseis#16 @mrosskopf
Of course! Done!
