### 2024-10-15 swiss-seismological-service/DUGseis#14 @mrosskopf
Thanks for pointing that out. For whatever reason, there was an old file within the folder with unchanged directories. I uploaded another new version, since I can not delete the old one.
