### 2024-10-17 swiss-seismological-service/DUGseis#14 @erexer
Great, its working for me now.
