### 2024-10-17 swiss-seismological-service/DUGseis#16 @erexer
I'm getting this error now:
```bash
raise ValueError(f"{error}\n{summary}") from None
      ValueError: invalid pyproject.toml config: `project.requires-python`.
      configuration error: `project.requires-python` must be pep508-versionspec
```
