### 2024-10-17 swiss-seismological-service/DUGseis#11 @erexer
Looks great!
