### 2024-10-18 swiss-seismological-service/DUGseis#16 @mrosskopf
Fixed!
