### 2024-10-21 swiss-seismological-service/DUGseis#16 @erexer
You're doing a great job, but this review is taking a lot of time because I'm testing/debugging your code, which is not the role of a reviewer. The `pyproject.toml` is still incorrect, let me know when it's fixed and I'll continue the review.
