+++ pyproject.toml
@@ -20,7 +20,7 @@ authors = [
 name = "DUG-Seis"
 version = "0.3"
 description = "Processing, data management, and visualization of micro-seismic data."
-requires-python = "==3.11.*"
+requires-python = ">=3.11"
 license = { text = "LGPL" }
 keywords = ["induced seismicity", "seismology"]
 classifiers = [
@@ -29,9 +29,8 @@ classifiers = [
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Topic :: Scientific/Engineering",
 ]
 dependencies = [
