### 2024-10-24 swiss-seismological-service/DUGseis#16 @krischer
@erexer Thanks a lot for all your work in the review.

We just relaxed the version constraint a bit [here](https://github.com/swiss-seismological-service/DUGseis/commit/4e6e6661779e27891f3374c0d0526c57be6ee1fa). Instead of being hard-coded it now only requires a Python version of at least 3.11.

It will currently work with 3.11 and 3.12 (and I just tested this with two clean environments); 3.13 will currently not work due `numba` still requiring an update (they typically take up to one year after a new release) but as soon as `numba` updates, it should work with our package.

In case it still does not work, please let us know the error you encounter.

All the best!
