### 2024-10-24 swiss-seismological-service/DUGseis#16 @krischer
One could of course argue that we should also fix the Python version to be smaller than 3.13 but I'd typically go with the advice in this blog post to not do that: https://iscinumpy.dev/post/bound-version-constraints/#pinning-the-python-version-is-special

Once all the dependencies work, DUGseis will likely also work with 3.13 without requiring a code change.
