### 2024-10-24 swiss-seismological-service/DUGseis#16 @erexer
Thanks for sharing that article, that's great to know! The install/versioning looks good to me.
