### 2024-10-24 swiss-seismological-service/DUGseis#16 @erexer
Can you change the version listed in the [docs](https://dugseis.readthedocs.io/en/latest/installation.html).
