### 2024-10-24 e6e9307
updated python version in documentation
