### 2024-10-24 swiss-seismological-service/DUGseis#16 @mrosskopf
I updated the version in the documentation to python >= 3.11.
