### 2024-10-24 #6768 @mrosskopf
Hi all, I just wanted to inform you that I will be on holidays until mid-November and won’t be able to react. I hope that the last part of the installation is now covered as well and that there are no other issues. It would be great to have the paper accepted with doi as soon as possible, since a handful of papers successfully using the code are very close to publication.
