### 2024-10-24 #6768 @crvernon
@jessepisel and @erexer - could you provide a summary of what is left from your side of things please?  Thank you!
