### 2024-11-16 #6768 @crvernon
👋  NAME - we are almost there!

- [ ] Please fix the invalid DOI in the above.  If that paper is still in review, please remove it from being sited.
- [ ] LINE 99:  Shouldn't "DUGSeis" be written "DUGseis" ?
- [ ] LINE 121:  The "p" in Python should be capitalized.  You can ensure formatting in your bib file by using curly brackets around what you wish to maintain the formatting of.
- [ ] LINE 129:  "gempa" should be all caps.
- [ ] LINE 140-141:  locational information needs captialization
- [ ] LINE 150:  need capitalization

Please correct these.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (make sure the list is correct and ordered as in the paper). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version here
