### 2024-11-19 #6768 @mrosskopf
Hi @crvernon, thank you! I corrected the points above and added the code to zenodo which has the DOI:
https://doi.org/10.5281/zenodo.14187295
