### 2024-11-22 #6768 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
