### 2024-11-22 #6768 @crvernon
🥳 Congratulations on your new publication @mrosskopf! Many thanks to @trichter, @erexer, and @jessepisel for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
